pyproject.toml
[tool.poetry]
name = "cryptodude"
version = "0.1.0"
description = ""
authors = ["Bogdan Abarikhin <bogdan_abarihin@mail.ru>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.75.2"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
cryptography = "^36.0.2"
passlib = "^1.7.4"
python-multipart = "^0.0.5"
uvicorn = "^0.17.6"
bcrypt = "^3.2.0"
environs = "^9.5.0"
black = "^22.3.0"
SQLAlchemy = "^1.4.35"
psycopg2-binary = "^2.9.3"

[tool.poetry.dev-dependencies]
pre-commit = "^2.15.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
